services/dashboard/requirements.txt
streamlit==1.36.0
pymongo==4.8.0
py2neo==2021.2.4
networkx==3.3
matplotlib==3.8.4
python-dotenv==1.0.1
